rfc5965: ignore bogus rfc822 dates arrival or reception-date headers
The rfc includes a bogus date in one of its examples, which the
underlying chrono library produces an error for.

Since we don't loudly report failures in parsing incoming reports,
this sort of error is confusing for the user.

These date fields are optional, so we can convert the error to
None without harming the overall outcome.