kumod: avoid panic when main policy file has syntax error
The log-accounting-db-on-flush logic was getting tripped up
in the case that you have a fatal syntax error, such as a
construct like `local foo = {` with no closing brace.

It would try to flush the db before shutting down, but we
hadn't yet initialized the root LifeCycle object, so it
would panic before we got a chance to print the details
of the syntax error.

This commit makes the flush logic recognize when we haven't
started a lifecycle and simply skip in that case, because
we couldn't possibly have passed traffic in that situation.