kumod: avoid spamming "shutting down" messages to log on shutdown
Depending on timing and overall system state, it is possible that we
could log a large number of "shutting down" errors to the log when a
shutdown is initiated.  Those errors are not useful to log to the
diagnostic, so we'd like to suppress them.

We had a couple of places doing some ad-hoc string comparisons for this
purpose, but it wasn't totally foolproof.

Neither is this commit, but it does formalize the shutting down error
state a bit better, and add an additional place where we can opt not to
log the associated error.