logging: introduce a log_disposition event hook
The higher level goal is to facilitate generation of RFC 3464
messages in response to delivery failures.

The first step is to introduce this synchronous (wrt. message processing
flow) event hook that will allow the message content to be optionally
captured by the hook implementation.  These need to be synchronous in
this way, otherwise a terminal dispoisition (eg: permanent failure) may
decide to remove the message from the spool concurrent with the log
processing.  That concurrency concern doesn't exist for the existing
logger implementations because they make a point of capturing all
information from the message prior to enqueuing the data to the logger.

Later commits will provide some convenience functions for bounce message
generation based upon that state.