mailparsing: accept commonly misphrased mime parameter header encoding
What a mess!

RFC 2047 defines an encoded-word as a way to indicate that some
text has 8-bit content and a specific charset, allowing for inline
transfer encoding to be applied to it.

It expressly says:

An 'encoded-word' MUST NOT appear within a 'quoted-string'.
An 'encoded-word' MUST NOT be used in parameter of a MIME
Content-Type or Content-Disposition field, or in any structured
field body except within a 'comment' or 'phrase'.

Today we encountered this header:

`Content-Disposition: attachment; filename="=?UTF-8?B?5pel5pys6Kqe44Gu5re75LuY?="`

which violates both of those restrictions.

RFC 2231 is a different RFC that specifies how to encode these sorts of
headers correctly.

Hunting around a bit on the internet, it appears that someone started to
use the bogus form, perhaps as a transitional workaround, or perhaps by
mistake, and it has become something of a de-facto accepted standard to
the point that gmail and fastmail both will generate this header
construction when attaching messages.

So we need to be able to parse these out.

This commit introduces a special case for this as the last step in the
MimeParameters::get method that will try to decode the retrieved value
as an encoded-word, and if successful, return that decoded value.