mailparsing: make Rfc2045Info::new infallible
If it fails, then the overall message parse attempt fails out,
but we want to allow operating on the borked message at some
basic level.

This commit adjusts the logic to assume reasonable fallback
values in case the various `Content-XXX` header parsing fails,
and to defer eg: detection of an unsupported charset until
the point where the content conversion is attempted.

The invalid headers are recorded as a new INVALID_MIME_HEADERS
conformance flag which causes check_fix_conformance to decide
that the message needs to be rebuilt.  The rebuild process
will come up with its own Content-XXX headers for the rebuilt
message through its existing logic.