mailparsing: preserve supplemental Content-Type parameters on rebuild
When performing a rebuild, we'd build completely new RFC 2045/RFC 2047
mime headers.

However, sometimes there are "load bearing" additional parameters in
the source part that are important to preserve; for example, calendar
invites include a `method=REQUEST` parameter that is important to
pass through in order for the invitation to correctly interface with
the recipients calendar.

This commit adjusts the rebuild logic to merge in any parameters
that we didn't set in the basic rebuild.