fixup test build
This is weird; I can't reproduce the CI failure locally, even with an
updated version of rust, but the diagnostic in CI points to
https://github.com/rust-lang/rust/issues/131114 which suggests that
str::from_ut8 is unstable, but buried within it is the mention that
std::str::from_utf8 is OK to use.

Let's see if this does the job?