smtp_server: improve cleanup for partially failed transaction
If we already saved one or more messages that were split out
of a multi-recipient transaction, and then the subsequent save
fails, we would previously leave those earlier messages saved
in the spool and deliver them despite reporting a failure back
to the client.

This commit removes them from the spool and logs a Bounce record
for them.

This is still not an ideal outcome, but it will at least not
send a message for which we haven't claimed responsibility.