templates: prepare to support multiple dialects of templates
This is primarily a refactoring of kumo-template to conceptually
support the idea of having multiple dialects, even though it
only defines a single dialect.

It prepares the types/interface for that purpose.

It also pulls in the handlebars crate, but doesn't use anything
from it yet; that'll be in a follow up commit.

refs: https://github.com/KumoCorp/kumomta/issues/446