docs: add breaking change note about RCPT TO 552 status
in 18e55a6e91d9cd132a712ee26932612815ff828e I noticed that we should be
handling 552 as a transient failure, but I neglected to mention this
change in the changelog because it seemed obscure and "couldn't" hurt
anyone.

I overlooked that we have client response rewriting and of course this
change did impact one of our sponsors that rewrites certain transient
452 responses to 552 responses.

So, let's add a note explaining this to see if we can prevent others
from hitting it in the future.