smtp_dispatcher: fix interpretation of 452 for single-recip messages
The multi-recipient send logic introduced a regression in `main` that
could cause messages like "452 4.1.1 mailbox full" to be considered
immediately retryable always, because we thought it meant "too many
recipients" in this batch.

This commit fixes up the immediate retry logic to only apply when we
have a batch size > 1, and we try to be a bit more discerning with the
enhanced status code when it is present.

This issue isn't being mentioned in the changelog as a breaking change,
as the breakage is vs. an earlier `dev` build and not a prior stable
release.