config: refactor callback calling
Centralize the logic of walking and calling the registered handlers
so that we don't have 3 or more slightly different versions of it.

Introduce a CallbackDisposition type that can be used to drive
the default handling with more nuance; this will be useful
in a later commit where we need to distinguish between an
explicit nil being returned by a handler, and no handler
being defined.